refactor(frontend): Move API response types to todoApi.ts
