+++ frontend/app/types.ts
@@ -9,17 +9,3 @@ export interface Todo {
 
 export type CreateTodoDto = Pick<Todo, "title" | "description">;
 export type UpdateTodoDto = Partial<Pick<Todo, "title" | "description">>;
-
-export interface PaginationMeta {
-  totalCount: number;
-  totalPage: number;
-  currentPage: number;
-  itemsPerPage: number;
-  hasNext: boolean;
-  hasPrevious: boolean;
-}
-
-export interface GetTodoListResponse {
-  todo: Todo[];
-  meta: PaginationMeta;
-}
